Remove PABotBase1
