--- SerialPrograms/Source/Controllers/SerialPABotBase/Connection/BotBase.cpp
@@ -1,18 +1,0 @@
-/*  Pokemon Automation Bot-Base Interface
- *
- *  From: https://github.com/PokemonAutomation/
- *
- */
-
-#include "BotBaseMessage.h"
-#include "BotBase.h"
-
-//#include <iostream>
-//using std::cout;
-//using std::endl;
-
-namespace PokemonAutomation{
-
-
-
-}
