--- SerialPrograms/Source/Controllers/SerialPABotBase/SerialPABotBase_Descriptor.cpp
@@ -1,91 +1,0 @@
-/*  Serial Port (PABotBase) Interface
- *
- *  From: https://github.com/PokemonAutomation/
- *
- */
-
-#include "Common/Cpp/Json/JsonValue.h"
-#include "Controllers/ControllerTypeStrings.h"
-#include "SerialPABotBase_Descriptor.h"
-
-#include "NintendoSwitch/Controllers/SerialPABotBase/NintendoSwitch_SerialPABotBase_WiredController.h"
-
-//#include <iostream>
-//using std::cout;
-//using std::endl;
-
-namespace PokemonAutomation{
-
-template class InterfaceType_t<SerialPABotBase::SerialPABotBase_Descriptor>;
-
-namespace SerialPABotBase{
-
-
-
-
-bool SerialPABotBase_Descriptor::operator==(const ControllerDescriptor& x) const{
-    if (typeid(*this) != typeid(x)){
-        return false;
-    }
-    return m_name == static_cast<const SerialPABotBase_Descriptor&>(x).m_name;
-}
-
-
-std::string SerialPABotBase_Descriptor::display_name() const{
-    if (m_name.empty()){
-        return "(none)";
-    }
-    return m_name;
-}
-void SerialPABotBase_Descriptor::load_json(const JsonValue& json){
-    const std::string* name = json.to_string();
-    if (name == nullptr || name->empty()){
-        return;
-    }
-    m_name = *name;
-}
-JsonValue SerialPABotBase_Descriptor::to_json() const{
-    return m_name;
-}
-
-std::unique_ptr<ControllerConnection> SerialPABotBase_Descriptor::open_connection(Logger& logger) const{
-    if (m_name.empty()){
-        return nullptr;
-    }
-    return std::unique_ptr<ControllerConnection>(
-        new SerialPABotBase_Connection(logger, m_name)
-    );
-}
-std::unique_ptr<AbstractController> SerialPABotBase_Descriptor::make_controller(
-    Logger& logger,
-    ControllerConnection& connection,
-    ControllerType controller_type
-) const{
-    switch (controller_type){
-    case ControllerType::NintendoSwitch_WiredController:
-    case ControllerType::NintendoSwitch2_WiredController:
-        return std::unique_ptr<AbstractController>(
-            new PokemonAutomation::NintendoSwitch::SerialPABotBase_WiredController(
-                logger,
-                static_cast<SerialPABotBase_Connection&>(connection),
-                controller_type
-            )
-        );
-
-    default:;
-    }
-
-    logger.log(
-        std::string("Unsupported Controller Type: ") + CONTROLLER_TYPE_STRINGS.get_string(controller_type),
-        COLOR_RED
-    );
-    return nullptr;
-}
-
-
-
-
-
-
-}
-}
