+++ SerialPrograms/Source/NintendoSwitch/Controllers/PABotBase2/NintendoSwitch_PABotBase2_OemController.cpp
@@ -6,7 +6,7 @@
 
 #include "Common/PABotBase2/Controllers/PABotBase2_Controller_NS1_OemController.h"
 #include "Common/Cpp/ColoredText.h"
-#include "Controllers/SerialPABotBase/SerialPABotBase.h"
+#include "Controllers/SerialPort/SerialPABotBase.h"
 #include "NintendoSwitch/NintendoSwitch_Settings.h"
 #include "NintendoSwitch_PABotBase2_OemController.h"
