--- SerialPrograms/Source/NintendoSwitch/Controllers/SerialPABotBase/NintendoSwitch_SerialPABotBase_Controller.h
@@ -1,59 +1,0 @@
-/*  SerialPABotBase: Controller
- *
- *  From: https://github.com/PokemonAutomation/
- *
- */
-
-#ifndef PokemonAutomation_NintendoSwitch_SerialPABotBase_Controller_H
-#define PokemonAutomation_NintendoSwitch_SerialPABotBase_Controller_H
-
-#include "Controllers/SerialPABotBase/Connection/BotBase.h"
-#include "Controllers/SerialPABotBase/SerialPABotBase_Connection.h"
-#include "NintendoSwitch/Controllers/NintendoSwitch_ControllerWithScheduler.h"
-
-namespace PokemonAutomation{
-namespace NintendoSwitch{
-
-
-class SerialPABotBase_Controller : public ControllerWithScheduler{
-public:
-    SerialPABotBase_Controller(
-        Logger& logger,
-        RecursiveThrottler& logging_throttler,
-        ControllerType controller_type,
-        SerialPABotBase::SerialPABotBase_Connection& connection
-    );
-
-    uint32_t protocol_version() const{
-        return m_handle.protocol_version();
-    }
-
-    void stop_with_error(std::string error_message) noexcept;
-
-    bool is_ready() const;
-    std::string error_string() const;
-
-
-public:
-    void cancel_all_commands();
-    void replace_on_next_command();
-
-    void wait_for_all(Cancellable* cancellable);
-
-
-protected:
-    //  These are set on construction and never changed again. So it is safe to
-    //  access these asynchronously.
-    SerialPABotBase::SerialPABotBase_Connection& m_handle;
-    BotBaseController* m_serial;
-
-    mutable SpinLock m_error_lock;
-    std::string m_error_string;
-};
-
-
-
-
-}
-}
-#endif
