--- SerialPrograms/Source/NintendoSwitch/Controllers/SerialPABotBase/NintendoSwitch_SerialPABotBase_WiredController.h
@@ -1,242 +1,0 @@
-/*  SerialPABotBase: Wired Controller (Switch 1)
- *
- *  From: https://github.com/PokemonAutomation/
- *
- */
-
-#ifndef PokemonAutomation_NintendoSwitch_SerialPABotBase_WiredControllerNS1_H
-#define PokemonAutomation_NintendoSwitch_SerialPABotBase_WiredControllerNS1_H
-
-#include "Controllers/ControllerStatusThread.h"
-#include "NintendoSwitch/NintendoSwitch_Settings.h"
-#include "NintendoSwitch/Controllers/Procon/NintendoSwitch_ProController.h"
-#include "NintendoSwitch_SerialPABotBase_Controller.h"
-
-namespace PokemonAutomation{
-namespace NintendoSwitch{
-
-
-
-
-class SerialPABotBase_WiredController final :
-    public ProController,
-    public SerialPABotBase_Controller,
-    private ControllerStatusThreadCallback
-{
-public:
-    using ContextType = ProControllerContext;
-
-
-public:
-    SerialPABotBase_WiredController(
-        Logger& logger,
-        SerialPABotBase::SerialPABotBase_Connection& connection,
-        ControllerType controller_type
-    );
-    ~SerialPABotBase_WiredController();
-    void stop();
-
-    virtual Logger& logger() override{
-        return m_logger;
-    }
-    virtual bool is_ready() const override{
-        return SerialPABotBase_Controller::is_ready();
-    }
-
-
-public:
-    virtual ControllerPerformanceClass performance_class() const override{
-        return ControllerPerformanceClass::SerialPABotBase_Wired;
-    }
-    virtual Milliseconds ticksize() const override{
-        return Milliseconds(0);
-    }
-    virtual Milliseconds cooldown() const override{
-        return Milliseconds(8);
-    }
-    virtual Milliseconds timing_variation() const override{
-        return ConsoleSettings::instance().TIMING_OPTIONS.WIRED;
-    }
-    virtual bool atomic_multibutton() const override{
-        return true;
-    }
-
-
-public:
-    virtual void cancel_all_commands() override{
-        SerialPABotBase_Controller::cancel_all_commands();
-    }
-    virtual void replace_on_next_command() override{
-        SerialPABotBase_Controller::replace_on_next_command();
-    }
-
-    virtual void wait_for_all(Cancellable* cancellable) override{
-        SerialPABotBase_Controller::wait_for_all(cancellable);
-    }
-
-
-public:
-    //  Superscalar Commands (the "ssf" framework)
-
-    virtual void issue_barrier(Cancellable* cancellable) override{
-        ControllerWithScheduler::issue_barrier(cancellable);
-    }
-    virtual void issue_nop(Cancellable* cancellable, Milliseconds duration) override{
-        ControllerWithScheduler::issue_nop(cancellable, duration);
-    }
-    virtual void issue_buttons(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        Button button
-    ) override{
-        button &= VALID_PRO_CONTROLLER_BUTTONS;
-        ControllerWithScheduler::issue_buttons(cancellable, delay, hold, cooldown, button);
-    }
-    virtual void issue_dpad(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        DpadPosition position
-    ) override{
-        ControllerWithScheduler::issue_dpad(cancellable, delay, hold, cooldown, position);
-    }
-    virtual void issue_left_joystick(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        const JoystickPosition& position
-    ) override{
-        ControllerWithScheduler::issue_left_joystick(cancellable, delay, hold, cooldown, position);
-    }
-    virtual void issue_right_joystick(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        const JoystickPosition& position
-    ) override{
-        ControllerWithScheduler::issue_right_joystick(cancellable, delay, hold, cooldown, position);
-    }
-
-    virtual void issue_gyro_accel_x(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        int16_t value
-    ) override{
-        ControllerWithScheduler::issue_gyro_accel_x(cancellable, delay, hold, cooldown, value);
-    }
-    virtual void issue_gyro_accel_y(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        int16_t value
-    ) override{
-        ControllerWithScheduler::issue_gyro_accel_y(cancellable, delay, hold, cooldown, value);
-    }
-    virtual void issue_gyro_accel_z(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        int16_t value
-    ) override{
-        ControllerWithScheduler::issue_gyro_accel_z(cancellable, delay, hold, cooldown, value);
-    }
-    virtual void issue_gyro_rotate_x(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        int16_t value
-    ) override{
-        ControllerWithScheduler::issue_gyro_rotate_x(cancellable, delay, hold, cooldown, value);
-    }
-    virtual void issue_gyro_rotate_y(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        int16_t value
-    ) override{
-        ControllerWithScheduler::issue_gyro_rotate_y(cancellable, delay, hold, cooldown, value);
-    }
-    virtual void issue_gyro_rotate_z(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        int16_t value
-    ) override{
-        ControllerWithScheduler::issue_gyro_rotate_z(cancellable, delay, hold, cooldown, value);
-    }
-
-    virtual void issue_full_controller_state(
-        Cancellable* cancellable,
-        bool enable_logging,
-        Milliseconds duration,
-        Button button,
-        DpadPosition dpad,
-        const JoystickPosition& left_joystick,
-        const JoystickPosition& right_joystick
-    ) override{
-        ControllerWithScheduler::issue_full_controller_state(
-            cancellable,
-            enable_logging,
-            duration,
-            button,
-            dpad,
-            left_joystick,
-            right_joystick
-        );
-    }
-
-
-public:
-    //  High speed RPCs.
-
-    virtual void issue_mash_button(
-        Cancellable* cancellable,
-        Milliseconds duration,
-        Button button,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown
-    ) override{
-        button &= VALID_PRO_CONTROLLER_BUTTONS;
-        ControllerWithScheduler::issue_mash_button(cancellable, duration, button, delay, hold, cooldown);
-    }
-    virtual void issue_mash_button(
-        Cancellable* cancellable,
-        Milliseconds duration,
-        Button button0, Button button1
-    ) override{
-        button0 &= VALID_PRO_CONTROLLER_BUTTONS;
-        button1 &= VALID_PRO_CONTROLLER_BUTTONS;
-        ControllerWithScheduler::issue_mash_button(cancellable, duration, button0, button1);
-    }
-    virtual void issue_mash_AZs(
-        Cancellable* cancellable,
-        Milliseconds duration
-    ) override{
-        ControllerWithScheduler::issue_mash_AZs(cancellable, duration);
-    }
-    virtual void issue_system_scroll(
-        Cancellable* cancellable,
-        Milliseconds delay, Milliseconds hold, Milliseconds cooldown,
-        DpadPosition direction  //  Diagonals not allowed.
-    ) override{
-        ControllerWithScheduler::issue_system_scroll(cancellable, delay, hold, cooldown, direction);
-    }
-
-
-private:
-    virtual void update_status(Cancellable& cancellable) override;
-    virtual void stop_with_error(std::string message) override;
-
-
-private:
-    template <typename Type>
-    PA_FORCE_INLINE Type milliseconds_to_ticks_8ms(Type milliseconds){
-        return milliseconds / 8 + (milliseconds % 8 + 7) / 8;
-    }
-    virtual void execute_state(
-        Cancellable* cancellable,
-        const SuperscalarScheduler::ScheduleEntry& entry
-    ) override;
-
-
-private:
-    std::unique_ptr<ControllerStatusThread> m_status_thread;
-};
-
-
-
-
-}
-}
-#endif
